+++ test/production/app-dir/metadata-revalidate/app/manifest.ts
