+++ test/production/app-dir/metadata-revalidate/app/revalidate/og/opengraph-image.tsx
